refactor(api): split poll-vote.service.ts into services/poll-vote/ (api-sr1o)
