+++ apps/api/src/__tests__/services/poll-vote.service.update-delete-results.test.ts
@@ -87,8 +87,11 @@ vi.mock("drizzle-orm", async (importOriginal) => {
 // ── Imports after mocks ──────────────────────────────────────────────
 
 const { assertSystemOwnership } = await import("../../lib/system-ownership.js");
-const { updatePollVote, deletePollVote, getPollResults, castVote, listVotes, parsePollVoteQuery } =
-  await import("../../services/poll-vote.service.js");
+const { castVote } = await import("../../services/poll-vote/cast.js");
+const { listVotes, parsePollVoteQuery } = await import("../../services/poll-vote/list.js");
+const { updatePollVote } = await import("../../services/poll-vote/update.js");
+const { deletePollVote } = await import("../../services/poll-vote/archive.js");
+const { getPollResults } = await import("../../services/poll-vote/results.js");
 
 // ── Fixtures ─────────────────────────────────────────────────────────
