+++ apps/api/src/routes/polls/cast-vote.ts
@@ -9,7 +9,7 @@ import { parseJsonBody } from "../../lib/parse-json-body.js";
 import { envelope } from "../../lib/response.js";
 import { createIdempotencyMiddleware } from "../../middleware/idempotency.js";
 import { createCategoryRateLimiter } from "../../middleware/rate-limit.js";
-import { castVote } from "../../services/poll-vote.service.js";
+import { castVote } from "../../services/poll-vote/cast.js";
 
 import type { AuthEnv } from "../../lib/auth-context.js";
