+++ apps/api/src/routes/polls/update-vote.ts
@@ -7,7 +7,7 @@ import { parseIdParam, requireIdParam } from "../../lib/id-param.js";
 import { parseJsonBody } from "../../lib/parse-json-body.js";
 import { envelope } from "../../lib/response.js";
 import { createCategoryRateLimiter } from "../../middleware/rate-limit.js";
-import { updatePollVote } from "../../services/poll-vote.service.js";
+import { updatePollVote } from "../../services/poll-vote/update.js";
 
 import type { AuthEnv } from "../../lib/auth-context.js";
